refactor:#70 프리센트이벤트 유저 접속시간 판단하는 로직 추상화
