+++ src/modules/mandalart/hooks/use-realtime-presence-room.ts
@@ -5,6 +5,7 @@ import { useCurrentUserImage } from './use-current-user-image';
 import { getBrowserClient } from '@/shared/utils/supabase/browser-client';
 import { useBroadcastStore } from './use-broadcast-store';
 import { useReceiveBroadcastStore } from './use-receive-broadcast-store';
+import { checkUserJoinTime } from '../utils/check-user-join-time';
 
 const supabase = getBrowserClient();
 
@@ -21,14 +22,15 @@ export type RealtimeUser = {
  */
 export const useRealtimePresenceRoom = (roomName: string, username: string) => {
   // 현재 나의 이미지와 이름 받아오기
-  const currentUserImage = useCurrentUserImage();
-  const currentUserName = username;
   const [users, setUsers] = useState<Record<string, RealtimeUser>>({});
   const myJoinTime = useRef<number>(Date.now());
   const { receiveBroadcastStore } = useReceiveBroadcastStore();
   const formatBroadcastStorePayload = useBroadcastStore(
     (state) => state.formatBroadcastStorePayload
   );
+  const currentUserImage = useCurrentUserImage();
+  const currentUserName = username;
+
   useEffect(() => {
     const room = supabase.channel(roomName);
 
@@ -40,38 +42,24 @@ export const useRealtimePresenceRoom = (roomName: string, username: string) => {
         joinTime: number;
       }>();
 
-      // presence 상태를 사용자별 객체로 변환
-      // newState는 { userId: [{ name: '이름', image: 'URL' }] } 형태
-      // { userId: { name: '이름', image: 'URL' } } 형태로 변환
+      // { userId(랜덤UUID): { name: '이름', image: 'URL' } } 형태로 변환
       const formattingAllUsersInfo = Object.fromEntries(
         Object.entries(allUsersInfo).map(([key, values]) => [
           key,
           { name: values[0].name, image: values[0].image },
         ])
       ) as Record<string, RealtimeUser>;
+      setUsers(formattingAllUsersInfo);
 
-      //가장 먼저 들어온 유저의 접속시간 찾기
-
-      const usersJoinTimeArr = Object.values(allUsersInfo)
-        .filter((userInfo) => userInfo[0].joinTime)
-        .map((userInfo) => {
-          return userInfo[0].joinTime;
-        });
-      const firstUserJoinTime = Math.min(...usersJoinTimeArr);
-      const newUserJoinTime = Math.max(...usersJoinTimeArr);
-
-      //방금 들어온 사람이면 다른 유저한테 브로드캐스트스토어 보내달라고 요청보내기
-      if (
-        usersJoinTimeArr.length !== 1 &&
-        newUserJoinTime === myJoinTime.current
-      ) {
+      const { firstUserJoinTime, newUserJoinTime, isNewUser } =
+        checkUserJoinTime(allUsersInfo, myJoinTime.current);
+      if (isNewUser) {
         room.send({
           type: 'broadcast',
           event: 'request_broadcasts_store',
           payload: { firstUserJoinTime, newUserJoinTime },
         });
       }
-      setUsers(formattingAllUsersInfo);
     });
 
     //가장 먼저 들어온 유저면 브로드캐스트스토어 보내주기(request_broadcasts_store에 대한 응답을 보내줌)
